+++ docs/build
@@ -1,4 +1,4 @@
 #!/bin/bash -eu
 
-sphinx-apidoc -efT -o docs/_apidoc nro45_merge
+sphinx-apidoc -efT -o docs/_apidoc mao_merge_45m
 sphinx-build -a docs docs/_build
