+++ docs/conf.py
@@ -25,5 +25,5 @@ templates_path = ["_templates"]
 # options for HTML output
 html_theme = "pydata_sphinx_theme"
 html_theme_options = {
-    "github_url": "https://github.com/mao-wfs/nro45-merge/",
+    "github_url": "https://github.com/mao-wfs/mao-merge-45m/",
 }
