+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "nro45-merge"
+name = "mao-merge-45m"
 version = "0.1.0"
 description = "Merge MAO datasets observed with NRO 45m telescope into a single Zarr file"
 authors = ["Akio Taniguchi <taniguchi@a.phys.nagoya-u.ac.jp>"]
